Interactive Word Bank with Pop Ups
The Word Search activity now has a Word Bank that is a GridView of
TextViews. The user can press a Word in a TextView to see a pop up of
that will contain information specific to the word.
